Making activity with embeddingV2 as main launcher
