+++ packages/path_provider/example/android/gradle.properties
@@ -1,2 +1,4 @@
 org.gradle.jvmargs=-Xmx1536M
+android.enableJetifier=false
+android.useAndroidX=true
 android.enableR8=true
